+++ Anime/App.js
@@ -1,5 +1,50 @@
+import React, { useState, useEffect } from "react";
+import { View, ActivityIndicator, StyleSheet } from "react-native";
 import OnBoarding from "./app/components/Onboarding/OnBoarding.js";
+import HomeScreen from "./app/components/HomeScreen/HomeScreen.js";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+
+const Loading = () => {
+  return (
+    <View>
+      <ActivityIndicator size="large" />
+    </View>
+  );
+};
 
 export default function App() {
-  return <OnBoarding />;
+  const [loading, setLoading] = useState(true);
+  const [viewOnboarding, setViewOnboarding] = useState(false);
+
+  const checkOnboarding = async () => {
+    try {
+      const value = await AsyncStorage.getItem("@viewedOnboarding");
+
+      if (value !== null) {
+        setViewOnboarding(true);
+      }
+    } catch (err) {
+      console.log("Error @checkOnboarding: ", err);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    checkOnboarding();
+  }, []);
+
+  return (
+    <View style={styles.container}>
+      {loading ? <Loading /> : viewOnboarding ? <HomeScreen /> : <OnBoarding />}
+    </View>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});