+++ Anime/app/components/Onboarding/OnBoarding.js
@@ -1,9 +1,11 @@
 import { StyleSheet, View, FlatList, Animated } from "react-native";
 import React, { useState, useRef } from "react";
+import AsyncStorage from "@react-native-async-storage/async-storage";
 
 import slides from "../../../utils/slides";
 import OnBoardingItem from "./OnBoardingItem";
 import Paginator from "./Paginator";
+import NextButton from "../../NextButton/NextButton";
 
 export default OnBoarding = () => {
   const [currentIndex, setCurrentIndex] = useState(0);
@@ -15,6 +17,20 @@ export default OnBoarding = () => {
   }).current;
 
   const viewConfig = useRef({ viewAreaCoveragePercentThreshold: 50 }).current;
+
+  //To access the slides using NextButton
+  const scrollTo = async () => {
+    //Checking if there any slides left to go
+    if (currentIndex < slides.length - 1) {
+      slidesRef.current.scrollToIndex({ index: currentIndex + 1 });
+    } else {
+      try {
+        await AsyncStorage.setItem("@viewedOnboarding", "true");
+      } catch (err) {
+        console.log("Error @setItem: ", err);
+      }
+    }
+  };
 
   return (
     <View style={styles.container}>
@@ -40,6 +56,10 @@ export default OnBoarding = () => {
         />
       </View>
       <Paginator data={slides} scrollX={scrollX} />
+      <NextButton
+        scrollTo={scrollTo}
+        percentage={(currentIndex + 1) * (100 / slides.length)}
+      />
     </View>
   );
 };